Add support for keeping peer session certificates
wolfSSL can keep the peer certificate loaded in memory for an SSL/TLS
session, allowing it to be retrieved after the session has been
established using `wolfSSL_get_peer_certificate`. This code is enabled
if wolfSSL is built with the `KEEP_PEER_CERT` token defined, but this is
typically only enabled alongside other features that we may or may not
be interested in.

This change exposes the `KEEP_PEER_CERT` option as a feature flag,
explicitly defining it as either a command-line compiler flag for
configure script builds or as part of `wolfcrypt/settings.h` for SGX
builds.