Support debug build configurations
This adds support for building wolfSSL with debugging support. Debugging
support is enabled automatically if `wolfssl-src` is built in debug mode
(based on checking for the `debug_assertions` configuration flag), which
is typically the case when `wolfssl-src` is used in a `-sys` crate, but
this can also be overridden using the `Build::debug` method. In addition
to disabling compiler optimizations, the source tree is not deleted from
the crate's `OUT_DIR` after building so that the debugger can resolve
source lines successfully when this option is enabled.